requirements.txt
amqp==1.4.9
anyjson==0.3.3
bcrypt==3.2.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==2.2.2
django-celery==3.3.1
django-cors-headers==2.1.0
djangorestframework==3.12.2
docopt==0.6.2
drf-yasg2==1.19.4
ecdsa==0.16.1
et-xmlfile==1.0.1
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==3.0.37
MarkupSafe==1.1.1
numpy==1.19.4
openpyxl==3.0.5
packaging==20.7
pandas==1.1.5
paramiko==1.17.0
pipreqs==0.4.10
pycparser==2.20
pycrypto==2.6.1
PyMySQL==0.8.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-crontab==2.2.8
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
xlrd==1.2.0
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.2.0
APScheduler==3.6.3
pytest==6.2.2
selenium==3.141.0
django-db-connection-pool==1.0.7
SQLAlchemy==1.4.11
func-timeout==4.3.5
